Add pagination for books. Change int to uint in places sensitive to negative values. Add caching for books. And other minor changes

+++ pkg/models/pagination.go
@@ -7,26 +7,26 @@ import (
 )
 
 type Pagination struct {
-	Limit      int         `json:"limit,omitempty" query:"limit"`
-	Page       int         `json:"page,omitempty" query:"page"`
+	Limit      uint         `json:"limit,omitempty" query:"limit"`
+	Page       uint         `json:"page,omitempty" query:"page"`
 	Sort       string      `json:"sort,omitempty" query:"sort"`
-	TotalRows  int64       `json:"total_rows"`
-	TotalPages int         `json:"total_pages"`
+	TotalRows  uint64       `json:"total_rows"`
+	TotalPages uint         `json:"total_pages"`
 	Rows       any         `json:"rows"`
 }
 
-func (p *Pagination) GetOffset() int {
+func (p *Pagination) GetOffset() uint {
 	return (p.GetPage() - 1) * p.GetLimit()
 }
 
-func (p *Pagination) GetLimit() int {
+func (p *Pagination) GetLimit() uint {
 	if p.Limit == 0 {
 		p.Limit = 10
 	}
 	return p.Limit
 }
 
-func (p *Pagination) GetPage() int {
+func (p *Pagination) GetPage() uint {
 	if p.Page == 0 {
 		p.Page = 1
 	}
@@ -44,11 +44,11 @@ func Paginate(value any, pagination *Pagination, db *gorm.DB) func(db *gorm.DB) 
 	var totalRows int64
 	db.Model(value).Count(&totalRows)
 
-	pagination.TotalRows = totalRows
-	totalPages := int(math.Ceil(float64(totalRows) / float64(pagination.Limit)))
+	pagination.TotalRows = uint64(totalRows)
+	totalPages := uint(math.Ceil(float64(totalRows) / float64(pagination.Limit)))
 	pagination.TotalPages = totalPages
 
 	return func(db *gorm.DB) *gorm.DB {
-		return db.Offset(pagination.GetOffset()).Limit(pagination.GetLimit()).Order(pagination.GetSort())
+		return db.Offset(int(pagination.GetOffset())).Limit(int(pagination.GetLimit())).Order(pagination.GetSort())
 	}
 }
